Add copy plugin command
