+++ src/main/groovy/org/craftercms/cli/Main.groovy
@@ -16,14 +16,21 @@
 
 package org.craftercms.cli
 
-import org.craftercms.cli.commands.*
+import org.craftercms.cli.commands.AddEnvironment
+import org.craftercms.cli.commands.marketplace.CopyPlugin
+import org.craftercms.cli.commands.site.AddRemote
+import org.craftercms.cli.commands.site.CreateSite
+import org.craftercms.cli.commands.site.ListRemotes
+import org.craftercms.cli.commands.site.ListSites
+import org.craftercms.cli.commands.site.SyncFrom
+import org.craftercms.cli.commands.site.SyncTo
 import picocli.CommandLine
 
 @CommandLine.Command(
         name = 'crafter-cli', usageHelpAutoWidth = true,
         versionProvider = VersionProvider.class, mixinStandardHelpOptions = true,
         subcommands = [CommandLine.HelpCommand, AddEnvironment, AddRemote, CreateSite, ListRemotes, SyncFrom, SyncTo,
-                        ListSites])
+                        ListSites, CopyPlugin])
 class Main {
 
     static def main(args) {
